Flyttet Amelding-kode fra AaregClient til egen AmeldingService.
